refactor duplication/clone assets code into it's own manager class as well as closing asset modals after performing duplication
